Split attendees.test.ts into per-function files
The 951-line attendees.test.ts mixed ambiguously-named groups
("CRUD", "stats", "decryption") with function-specific ones. Split
into one file per public function under test/lib/db/attendees/ so
each file's scope matches exactly one exported API.

Moved two tests that exercised getEventWithCount / getAllEvents
attendee counts into events.test.ts where they belong.
